Restructure project to use LangChain (not working)
The broad storkes of the project are there, however, without a proper
database to store chats, meesages, roles, etc... it is not possible to
use LangChain properly. The project is functional, but the chatbot's
memory persists even through server restarts. Rest assured, this will
be fixed in the next few commits.